Suppress unavoidable warnings
Some code compiled with warnings. One warning was related to the strong type system in Java and how it interacts with object deserialisation. The other warning occurred multiple times and was due to the use of getChr() which is a deprecated method in htsjdk - however many versions of htsjdk still in regular use do not have the replacement method. These warning have therefore been suppressed.