Better scores for hom/hemi deletions
When normalised read depth is very low, the estimate of error is an overestimate. Particularly, when the normalised read depth is zero, that is a very strong signal of a homozygous or hemizygous deletion, and the current error estimation mechanism does not take this into account. This change uses the Poisson distribution to produce a more realistic estimate of normalised read depth error when the normalised read depth is lower than 0.4. This causes smaller hom/hemi deletions to be detected, and such deletions to be given a higher quality score.